+++ src/cmd/parallel.rs
@@ -628,11 +628,8 @@ Finally, preprocessing on each file can be done using two different methods:
 1. Using only xan subcommands with -P, --preprocess:
     $ xan parallel count -P \"search -s name John | slice -l 10\" file.csv
 
-2. Using a shell subcommand passed to \"$SHELL -c\" with -H, --shell-preprocess:
-    $ xan parallel count -H \"xan search -s name John | xan slice -l 10\" file.csv
-
-The second preprocessing option will of course not work in DOS-based shells and Powershell
-on Windows.
+2. Using a subcommand passed to \"$SHELL -c\" or \"cmd /C\" with -H, --shell-preprocess:
+    $ xan parallel count -H \"rg john | xan from -f ndjson\" data.ndjson
 
 Usage:
     xan parallel count [options] [<inputs>...]
@@ -655,7 +652,7 @@ Usage:
 parallel options:
     -P, --preprocess <op>        Preprocessing using only `xan` subcommands.
     -H, --shell-preprocess <op>  Preprocessing commands that will run directly in your
-                                 own shell using the -c flag. Will not work on windows.
+                                 own shell using the -c flag.
     --progress                   Display a progress bar for the parallel tasks. The
                                  per file/chunk bars will tick once per CSV row only
                                  AFTER pre-processing!
@@ -901,7 +898,11 @@ impl Args {
                 .delimiter(self.flag_delimiter)
                 .no_headers(self.flag_no_headers);
 
-            let shell = env::var("SHELL").map_err(|_| "$SHELL is not set!")?;
+            let shell = if cfg!(target_os = "windows") {
+                "cmd"
+            } else {
+                &env::var("SHELL").map_err(|_| "$SHELL is not set!")?
+            };
 
             let mut cmd = Command::new(shell);
             let mut children: Vec<Child> = Vec::new();
@@ -938,7 +939,14 @@ impl Args {
             let mut child = cmd
                 .stdout(Stdio::piped())
                 .stderr(Stdio::piped())
-                .args(["-c", preprocessing])
+                .args([
+                    if cfg!(target_os = "windows") {
+                        "/C"
+                    } else {
+                        "-c"
+                    },
+                    preprocessing,
+                ])
                 .spawn()?;
 
             let reader = Box::new(child.stdout.take().expect("cannot read child stdout"));