+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Adding `xan window -C/--along-columns`.
 * Adding `xan cat rows --raw, -P/--preprocess & -H/--shell-preprocess`.
 * Improving `xan select` DSL star selectors. You can now do stuff like `vec_*_count`, `*[1]`, `vec_*[1]` etc.
+* `xan p -H/--shell-preprocess` now works on Windows.
 * Adding native `zsh` completions (@apcamargo).
 
 *Fixes*